requirements.txt
requests>=2.32.0
python-dotenv>=1.1.0
numpy>=2.3.0